moved getShiftType to jobRepository
