+++ app/src/main/java/com/example/oddhours/utils/Constants.kt
@@ -8,4 +8,7 @@ object Constants {
     const val OVERNIGHT_SHIFT = 1
     const val DAY_SHIFT = 2
     const val INVALID_SHIFT_RANGE = 0
+
+    const val SATURDAY = 7
+    const val SUNDAY = 1
 }
